requirements.txt
streamlit==1.32.0
sentence-transformers==2.6.1
torch
PyPDF2==3.0.1
docx2txt==0.9
fpdf2==2.7.8
